Handle http/https traffic correctly
* as of version Pie of Android (API level 28) the
isCleartextTrafficPermitted() method on NetworkSecurityPolicy returns
false per default (before that it defaulted to true). This means that
insecure web traffic is not allowed and the WebView component will
obey this setting since API level 26

* there are two ways to make the isCleartextTrafficPermitted return
true:
  * in the manifest file include <application
    android:usesCleartextTraffic="true"
  * in a network security configuration, but in that case you can only
    specify cleartextTrafficPermitted="true" for an enumerated set of domains

* it is impossible to know in advance which URL's a user has configured
in his/her apps so using the network configuration file will lead to
insufficient sites being white listed

* currently the quality guidelines state that an app needs to have a
network security configuration file and not specifying any domains in
a domain-config (the location where cleartextTrafficPermitted is
specified) results in linting errors on this configuration

* therefore a network security configuration is included with a basic
set of domains (insufficient for a properly working application however)
and additionally the android:usesCleartextTraffic setting is used in the
manifest file

PLEASE NOTE: this is probably not in line with the quality guidelines but
needed to allow proper functioning of the app. There seems no other
solution to achieve this.

PLEASE NOTE: it is also a quality guideline to only use SSL. This is
properly meant for calls from the Android Java code over the internet to
backend services and not specific for code running inside a WebView but
the fact that WebView honors the isCleartextTrafficPermitted setting is
an indication of the direction that Google is heading with Android.
Moving the RSS feeds we use to HTTPS and enabling customers to use HTTPS
wherever possible seems a good idea.