Handle secure traffic setting gracefully
* although we have set the handling of network traffic to use insecure
protocols (ie http) the writing is on the wall that this will not be
allowed for long. If for some reason https traffic is required it is
unwise to let the app "freeze" on startup. This would happen when
insecure network traffic is not allowed and at startup the WebView is
set to go to http://play.playr.biz. Since we support https on playback
it seems wise to use this and make the URL that is used to initiate the
WebView "aware" of the secure network traffic setting.

* in the MainActivity a check is done on the runtime value of the
setting that dictates the use of insecure network traffic and that is
communicated to the playr_loader page using the url. The playr_loader
page checks the query string for this parameter and switches between
the http and https protocol accordingly.